+++ src/audiosource/sfxr/soloud_sfxr.cpp
@@ -155,7 +155,7 @@ namespace SoLoud
 			if (env_stage == 1)
 			{
 				if (env_length[1])
-					env_vol = 1.0f + pow(1.0f - (float)env_time / env_length[1], 1.0f)*2.0f*mParams.p_env_punch;
+					env_vol = 1.0f + (float)pow(1.0f - (float)env_time / env_length[1], 1.0f)*2.0f*mParams.p_env_punch;
 				else
 					env_vol = 0;
 			}
@@ -282,16 +282,16 @@ namespace SoLoud
 			// reset filter
 			fltp=0.0f;
 			fltdp=0.0f;
-			fltw=pow(mParams.p_lpf_freq, 3.0f)*0.1f;
+			fltw=(float)pow(mParams.p_lpf_freq, 3.0f)*0.1f;
 			fltw_d=1.0f+mParams.p_lpf_ramp*0.0001f;
-			fltdmp=5.0f/(1.0f+pow(mParams.p_lpf_resonance, 2.0f)*20.0f)*(0.01f+fltw);
+			fltdmp=5.0f/(1.0f+(float)pow(mParams.p_lpf_resonance, 2.0f)*20.0f)*(0.01f+fltw);
 			if(fltdmp>0.8f) fltdmp=0.8f;
 			fltphp=0.0f;
-			flthp=pow(mParams.p_hpf_freq, 2.0f)*0.1f;
+			flthp=(float)pow(mParams.p_hpf_freq, 2.0f)*0.1f;
 			flthp_d=(float)(1.0+mParams.p_hpf_ramp*0.0003f);
 			// reset vibrato
 			vib_phase=0.0f;
-			vib_speed=pow(mParams.p_vib_speed, 2.0f)*0.01f;
+			vib_speed=(float)pow(mParams.p_vib_speed, 2.0f)*0.01f;
 			vib_amp=mParams.p_vib_strength*0.5f;
 			// reset envelope
 			env_vol=0.0f;
@@ -301,9 +301,9 @@ namespace SoLoud
 			env_length[1]=(int)(mParams.p_env_sustain*mParams.p_env_sustain*100000.0f);
 			env_length[2]=(int)(mParams.p_env_decay*mParams.p_env_decay*100000.0f);
 
-			fphase=pow(mParams.p_pha_offset, 2.0f)*1020.0f;
+			fphase=(float)pow(mParams.p_pha_offset, 2.0f)*1020.0f;
 			if(mParams.p_pha_offset<0.0f) fphase=-fphase;
-			fdphase=pow(mParams.p_pha_ramp, 2.0f)*1.0f;
+			fdphase=(float)pow(mParams.p_pha_ramp, 2.0f)*1.0f;
 			if(mParams.p_pha_ramp<0.0f) fdphase=-fdphase;
 			iphase=abs((int)fphase);
 			ipp=0;
